% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix2latex.R
\name{matrix2latex}
\alias{matrix2latex}
\title{matrix2latex}
\usage{
matrix2latex(x, type = "pmatrix")
}
\arguments{
\item{x}{matrix}

\item{type}{latex matrix environment}
}
\description{
Format a matrix for latex.
}
