% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers2words.R
\name{numbers2words}
\alias{numbers2words}
\title{Convert integers to English words.}
\usage{
numbers2words(x)
}
\arguments{
\item{x}{integer to format.}
}
\description{
\code{numbers2words} spells out integers in English.
The code is lifted from Andy Teucher \url{https://github.com/ateucher/useful_code/blob/master/R/numbers2words.r},
who in turn stole it from John Fox.
It has been improved somewhat by AAK
}
\examples{
numbers2words(49968883)
numbers2words(c(85999000,54,540,5400,54000,540000))
numbers2words(1e13+3)

}
