% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{random}
\alias{random}
\alias{random.org}
\alias{urandom}
\alias{rngControl}
\alias{rngSeeds}
\title{Functions for generating and working with truly random integers.}
\usage{
random.org(n = 10, rnd = "new")

urandom(n = 10)

rngControl(expr, seed = NULL)

rngSeeds(n, seed = NULL)
}
\arguments{
\item{n}{Number of integers required.}

\item{rnd}{random.org parameter}

\item{expr}{Expression to be evaluated with RNG control.}

\item{seed}{RNG seed.}
}
\value{
integers suitable for use as RNG seeds
}
\description{
Functions for generating and working with truly random seeds.
}
\details{
\code{random.org} gets seeds from \url{random.org}.

\code{urandom} gets seeds locally from \file{/dev/urandom} on *nix systems.

\code{rngControl} is a function to control RNG for the evaluation of an expression.

\code{rngSeeds} generates RNG seeds using \code{\link{sample.int}}.
It is included for situations when neither \code{\link{random.org}} nor \code{\link{urandom}} is available.
}
\examples{

random.org(n=5)
seed <- urandom(n=1)
seeds <- rngSeeds(5,seed=seed)
set.seed(seed)
runif(5)
rngControl(runif(5),seed=seed[1])
rngControl(runif(5),seed=seed[1])
runif(5)
set.seed(seed)
runif(5)
runif(5)

}
\references{
\url{http://www.random.org}
}
\author{
Aaron A. King
}
