% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scinot.R
\name{scinot}
\alias{scinot}
\title{Scientific notation.}
\usage{
scinot(x, digits = 2, format = c("expression", "latex", "math"),
  simplify = FALSE)
}
\arguments{
\item{x}{number(s) to format.}

\item{digits}{number of significant digits in mantissa.}

\item{format}{format specification.
\code{format="expression"} results in an R expression.
\code{format="latex"} results in a latex expression.
\code{format="math"} is like "latex" but wraps the text in "$".}

\item{simplify}{logical.
If \code{simplify=TRUE}, then \eqn{1 \times 10^n}{1x10^n} is simplified to \eqn{10^n}{10^n}.}
}
\description{
Format using scientific notation.
}
\examples{

x <- c(0.0309595,8577676.441,10000)
scinot(x[2],4)
scinot(x[1],2,"latex")
sapply(x,scinot,digits=3,format='math')
scinot(x,digits=0,simplify=FALSE)
scinot(x,digits=0,simplify=TRUE)

}
\seealso{
\code{\link[scales]{scientific}}
}
\author{
Aaron A. King
}
