% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunnel.R, R/aaa.R
\name{tunnel}
\alias{tunnel}
\alias{startTunnel}
\alias{stopTunnel}
\alias{cleanup}
\title{ssh tunneling}
\usage{
startTunnel(
  port = NULL,
  remotehost = getOption("aakmisc.remotehost", NULL),
  user = getOption("aakmisc.user", NULL),
  sleep = 5
)

stopTunnel(..., pid = getOption("aakmisc.tunnelpid", NULL))
}
\arguments{
\item{port}{Port on which to connect to PostgreSQL database.  If NULL, a
random port number will be used.}

\item{remotehost}{Hostname of PostgreSQL server.
An ssh tunnel to this host will be created.}

\item{user}{Username to use in conneting to PostgreSQL database.  If NULL,
Sys.getenv("USER") will be used.}

\item{sleep}{Time in seconds to sleep after initiating the ssh tunnel.}

\item{...}{Additional arguments ignored.}

\item{pid}{ID of ssh tunnel process.
Set automatically by \code{startTunnel}.}
}
\description{
Secure-shell (ssh) tunnels for database access.
}
