% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db}
\alias{db}
\alias{writeDBTable}
\alias{getQuery}
\alias{getMLEs}
\alias{recMLEs}
\alias{recScript}
\alias{dropScript}
\alias{listScripts}
\alias{catScript}
\title{Interface with databases}
\usage{
writeDBTable(
  name,
  value,
  overwrite = FALSE,
  append = FALSE,
  row.names = FALSE,
  host = getOption("aakmisc.dbhost", "localhost"),
  dbname = getOption("aakmisc.dbname", NULL),
  port = getOption("aakmisc.port", 5432),
  user = getOption("aakmisc.user", NULL),
  ...
)

getQuery(
  statement,
  host = getOption("aakmisc.dbhost", "localhost"),
  dbname = getOption("aakmisc.dbname", NULL),
  port = getOption("aakmisc.port", 5432),
  user = getOption("aakmisc.user", NULL),
  ...
)

getMLEs(
  host = getOption("aakmisc.dbhost", "localhost"),
  dbname = getOption("aakmisc.dbname", NULL),
  port = getOption("aakmisc.port", 5432),
  user = getOption("aakmisc.user", NULL),
  ...
)

recMLEs(
  mle,
  host = getOption("aakmisc.dbhost", "localhost"),
  dbname = getOption("aakmisc.dbname", NULL),
  port = getOption("aakmisc.port", 5432),
  user = getOption("aakmisc.user", NULL),
  ...
)

recScript(
  files,
  host = getOption("aakmisc.dbhost", "localhost"),
  dbname = getOption("aakmisc.dbname", NULL),
  port = getOption("aakmisc.port", 5432),
  user = getOption("aakmisc.user", NULL),
  ...
)

dropScript(
  script,
  host = getOption("aakmisc.dbhost", "localhost"),
  dbname = getOption("aakmisc.dbname", NULL),
  port = getOption("aakmisc.port", 5432),
  user = getOption("aakmisc.user", NULL),
  ...
)

listScripts(
  host = getOption("aakmisc.dbhost", "localhost"),
  dbname = getOption("aakmisc.dbname", NULL),
  port = getOption("aakmisc.port", 5432),
  user = getOption("aakmisc.user", NULL),
  ...
)

catScript(
  script,
  file = "",
  host = getOption("aakmisc.dbhost", "localhost"),
  dbname = getOption("aakmisc.dbname", NULL),
  port = getOption("aakmisc.port", 5432),
  user = getOption("aakmisc.user", NULL),
  ...
)
}
\arguments{
\item{name, value}{Name and contents of table to create.}

\item{overwrite, append, row.names}{See \code{\link[DBI]{dbWriteTable}}.}

\item{host}{Hostname on which to connect to the PostgreSQL server.}

\item{dbname}{Name of PostgreSQL database.}

\item{port}{Port on which to connect to PostgreSQL database.  If NULL, a
random port number will be used.}

\item{user}{Username to use in conneting to PostgreSQL database.  If NULL,
Sys.getenv("USER") will be used.}

\item{\dots}{Additional arguments will be passed to
\code{\link[DBI]{dbConnect}}.}

\item{statement}{SQL statement passed to \code{\link[DBI]{dbGetQuery}}.}

\item{mle}{A data-frame of MLEs to be recorded.}

\item{files}{Files containing R scripts to be recorded.}

\item{script}{Name of script.}

\item{file}{File to which the script will be written.  See
\code{\link{cat}}.}
}
\description{
Interface with project databases.
}
\examples{
\dontrun{
startTunnel()
listScripts()
stopTunnel()
}
}
\author{
Aaron A. King
}
