% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnc.R
\name{trnc}
\alias{trnc}
\title{Truncation of plots.}
\usage{
trnc(x, range = c(0, 1), only.finite = TRUE)
}
\arguments{
\item{x}{Numeric vector of values to manipulate.}

\item{range}{Numeric vector of length two giving desired output range.}

\item{only.finite}{if TRUE (the default), will only modify finite values.}
}
\description{
Truncate to the specified window.
}
\details{
\code{trnc} is a function for truncating data to a specified window.
It is suitable for use in \code{scale_{x,y}_{continuous,discrete}}, for example.
}
\examples{

trnc(c(-1,0.5,1,2,NA))

}
\seealso{
\code{\link[scales]{censor}}
}
\author{
Aaron A. King \email{kingaa at umich dot edu}
}
